+++ packages/components/scripts/generateDSRComponents.ts
@@ -489,6 +489,10 @@ import { get${componentName}Css } from '${stylesBundleImportPath}';
           // remove any jsx since options are not visible in closed multi-select
           .replace(/<>\s*([\s\S]*)\s*<\/>/, '<></>')
           .replace(/this\.theme/, 'this.props.theme');
+      } else if (tagName === 'p-optgroup') {
+        newFileContent = newFileContent
+          // remove any jsx since optgroups are not visible in closed multi-select | select
+          .replace(/<>\s*([\s\S]*)\s*<\/>/, '<></>');
       } else if (tagName === 'p-select') {
         newFileContent = newFileContent
           // replace wrapper className